+++ s4e-web/src/app/views/settings/settings.module.ts
@@ -1,12 +1,13 @@
+import { BreadcrumbService } from './breadcrumb/breadcrumb.service';
 
 import { ModalModule } from './../../modal/modal.module';
-import { GROUP_FORM_MODAL_ID } from './groups/group-form/group-form-modal.model';
+import { BreadcrumbModule } from './breadcrumb/breadcrumb.module';
 import { SearchModule } from './../../components/search/search.module';
 import { AdminDashboardModule } from './admin-dashboard/admin-dashboard.module';
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {RouterModule} from '@angular/router';
-import {settingsRoutes} from './settings.routes';
+import { RouterModule, ActivatedRoute, Router } from '@angular/router';
+import {settingsRoutes, SETTINGS_PATH} from './settings.routes';
 import {SettingsComponent} from './settings.component';
 import {ProfileModule} from './profile/profile.module';
 import {InstitutionProfileModule} from './intitution-profile/institution-profile.module';
@@ -35,8 +36,13 @@ import { NgxUiLoaderModule } from 'ngx-ui-loader';
     RouterModule.forChild(settingsRoutes),
     ManageInstitutionsModule,
     SearchModule,
-    ModalModule
+    ModalModule,
+    BreadcrumbModule
   ]
 })
 export class SettingsModule {
+  constructor(private _breadcrumbService: BreadcrumbService) {
+    const childrenRoutes = settingsRoutes.find(route => route.path === SETTINGS_PATH).children;
+    this._breadcrumbService.registerRoutes(childrenRoutes, SETTINGS_PATH);
+  }
 }