+++ s4e-web/src/app/views/settings/settings.routes.ts
@@ -1,40 +1,97 @@
-import { ChangePasswordComponent } from './profile/change-password/change-password.component';
+import { GroupListComponent } from './groups/group-list/group-list.component';
 import { InstitutionFormComponent } from './manage-institutions/institution-form/institution-form.component';
-import { AdminDashboardComponent } from './admin-dashboard/admin-dashboard.component';
+import { ChangePasswordComponent } from './profile/change-password/change-password.component';
+import { IBreadcrumb } from './breadcrumb/breadcrumb.model';
 import { Routes} from '@angular/router';
 import {ProfileComponent} from './profile/profile.component';
 import {SettingsComponent} from './settings.component';
-import {GroupListComponent} from './groups/group-list/group-list.component';
-import {InstitutionProfileComponent} from './intitution-profile/institution-profile.component';
-import {PersonListComponent} from './people/person-list/person-list.component';
 import {IsLoggedIn} from '../../utils/auth-guard/auth-guard.service';
-import {PersonFormComponent} from './people/person-form/person-form.component';
-import {GroupFormComponent} from './groups/group-form/group-form.component';
-import {DashboardComponent} from './dashboard/dashboard.component';
-import {IsManagerGuard} from './guards/is-manager/is-manager.guard';
-import { adminDashboardMatcher, managerDashboardMatcher } from 'src/app/utils';
-import {InstitutionListComponent} from './manage-institutions/institution-list/institution-list.component';
+import { adminDashboardMatcher, managerDashboardMatcher } from './dashboards.routes';
+import { AdminDashboardComponent } from './admin-dashboard/admin-dashboard.component';
+import { DashboardComponent } from './dashboard/dashboard.component';
+import { PersonListComponent } from './people/person-list/person-list.component';
+import { IsManagerGuard } from './guards/is-manager/is-manager.guard';
+import { InstitutionListComponent } from './manage-institutions/institution-list/institution-list.component';
+import { InstitutionProfileComponent } from './intitution-profile/institution-profile.component';
 
+export const userSettingsRoutes: Routes = [
+
+];
+
+export const SETTINGS_PATH = 'settings';
 export const settingsRoutes: Routes = [
   {
-    path: 'settings',
+    path: SETTINGS_PATH,
     component: SettingsComponent,
     canActivate: [IsLoggedIn],
     children: [
       {
         matcher: adminDashboardMatcher,
-        component: AdminDashboardComponent
+        component: AdminDashboardComponent,
+        data: {
+          isAdminDashboard: true,
+          breadcrumbs: [
+            {
+              label: 'Tablica administratora',
+              url: 'dashboard'
+            }
+          ]
+        }
       },
       {
         matcher: managerDashboardMatcher,
-        component: DashboardComponent
+        component: DashboardComponent,
+        data: {
+          isAdminDashboard: false,
+          breadcrumbs: [
+            {
+              label: 'Tablica menedżera',
+              url: 'dashboard'
+            }
+          ]
+        }
       },
+
+      {
+        path: 'institutions',
+        component: InstitutionListComponent,
+        canActivate: [IsManagerGuard],
+        data: {
+          breadcrumbs: [
+            {
+              label: 'Instytucje'
+            }
+          ]
+        }
+      },
+      {
+        path: 'institution',
+        component: InstitutionProfileComponent,
+        canActivate: [IsManagerGuard],
+        data: {
+          breadcrumbs: [
+            {
+              label: 'Profil instytucji'
+            }
+          ]
+        }
+      },
+
       {
         path: 'add-institution',
         component: InstitutionFormComponent,
         canActivate: [IsManagerGuard],
         data: {
-          isEditMode: false
+          isEditMode: false,
+          breadcrumbs: [
+            {
+              label: 'Lista instytucji',
+              url: 'institutions'
+            },
+            {
+              label: 'Dodaj instytucję'
+            }
+          ]
         }
       },
       {
@@ -42,47 +99,68 @@ export const settingsRoutes: Routes = [
         component: InstitutionFormComponent,
         canActivate: [IsManagerGuard],
         data: {
-          isEditMode: true
+          isEditMode: true,
+          breadcrumbs: [
+            {
+              label: 'Profil instytucji',
+              url: 'institution'
+            },
+            {
+              label: 'Edytuj instytucję',
+            }
+          ]
         }
       },
-      {
-        path: 'profile',
-        component: ProfileComponent
-      },
-      {
-        path: 'change-password',
-        component: ChangePasswordComponent
-      },
+
       {
         path: 'groups',
         component: GroupListComponent,
-        canActivate: [IsManagerGuard]
+        canActivate: [IsManagerGuard],
+        data: {
+          breadcrumbs: [
+            {
+              label: 'Grupy'
+            }
+          ]
+        }
       },
-      {
-        path: 'groups/:groupSlug',
-        component: GroupFormComponent,
-        canActivate: [IsManagerGuard]
-      },
-      {
-        path: 'institution',
-        component: InstitutionProfileComponent,
-        canActivate: [IsManagerGuard]
-      },
-      {
-        path: 'institutions',
-        component: InstitutionListComponent,
-        canActivate: [IsManagerGuard]
-      },
+
       {
         path: 'people',
         component: PersonListComponent,
-        canActivate: [IsManagerGuard]
+        canActivate: [IsManagerGuard],
+        data: {
+          breadcrumbs: [
+            {
+              label: 'Ludzie'
+            }
+          ]
+        }
+      },
+
+      {
+        path: 'profile',
+        component: ProfileComponent,
+        data: {
+          breadcrumbs: [
+            {
+              label: 'Mój profil'
+            }
+          ]
+        }
       },
       {
-        path: 'people/:email',
-        component: PersonFormComponent,
-        canActivate: [IsManagerGuard]
+        path: 'change-password',
+        component: ChangePasswordComponent,
+        data: {
+          breadcrumbs: [
+            {
+              label: 'Zmiana hasła'
+            }
+          ]
+        }
       },
+
       {
         path: '**',
         redirectTo: 'profile',